edit diffusion_step1A to use edited bvals file if it exists, in case vols had to be manually removed

+++ analysis_diffusion_step1A_bedpostx.sh
@@ -11,5 +11,7 @@ DIFFDIR=${ANALYSISDIR}/${MYSUB}/diffusion
 cp $0 ${ANALYSISDIR}/${MYSUB}/.
 
 cp ${DIFFDIR}/data.eddy_rotated_bvecs ${DIFFDIR}/bvecs
-cp ${SCRIPTSDIR}/bvals ${DIFFDIR}/.
+if [ ! -f ${DIFFDIR}/bvals ]; then 
+  cp ${SCRIPTSDIR}/bvals ${DIFFDIR}/.
+fi
 bedpostx ${DIFFDIR}
